requirement.txt
Django==1.9
django-debug-toolbar==1.9.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
mysqlclient==1.3.13
pycodestyle==2.4.0
pytz==2018.5
sqlparse==0.2.4
